+++ praw/models/__init__.py
@@ -7,7 +7,7 @@ from .list.redditor import RedditorList
 from .list.trophy import TrophyList
 from .listing.domain import DomainListing
 from .listing.generator import ListingGenerator
-from .listing.listing import Listing
+from .listing.listing import Listing, ModeratorListing
 from .mod_action import ModAction
 from .preferences import Preferences
 from .reddit.collections import Collection