+++ CHANGES.rst
@@ -6,10 +6,10 @@ Unreleased
 
 **Added**
 
-* Add method :meth:`~.Subreddits.premium` to reflect the naming change in
-  Reddit's API.
+* Add method :meth:`~.Subreddits.premium` to reflect the naming change in Reddit's API.
 * Ability to submit image galleries with :meth:`.submit_gallery`.
 * Ability to specify modmail mute duration.
+* Add method :meth:`.invited` to get invited moderators of a subreddit.
 
 **Changed**